Use sphinx-prompt extension
The sphinx-prompt extension provides nicer markup for command line
prompt instructions. It is ensures that the prompt character (`$`
by default) is not selectable when a user copies and pastes commands.
This improves UX.

See https://github.com/sbrunner/sphinx-prompt

Since we're using sphinx extensions, I've make it easier to install
dependencies with a requirements.txt. Added instructions to README.