Create a Development Workflow guide
This guide is meant to be sufficient for a new hire to get started with
the stack. It is intentionally light in details and special cases in
many circumstances because I assume there will be additional in-depth
guides written in the future that I can link to.

The guide focuses entirely on the lsstsw workflow since it was advocated
as the best general workflow by the team.
https://community.lsst.org/t/what-development-workflow-should-be-advocated-w-poll/297?u=jsick

A separate guide should be written around using the shared stack on the
NCSA cluster.

This guide is an expansion of what Russell Owen wrote at
https://community.lsst.org/t/developing-software-with-the-dm-stack/126?u=jsick

For DM-4016.