Add reStructuredText style guide.
This is an opinionated guide to using reST for LSST documentation.
This guide is needed so that we can guide developers towards the subset
and extensions of reST used by LSST since reST is a vast ecosystem with
lots of ways of marking up the same things, or allowing for types of
markup that we'd rather not have in our docs.

Incorporates improvements from Russell Owen's review:

- Put practical information up-front, rather than being didactic, moved
  formatting guidelines to the last section since they are only a backup
  to idiomatic use of reST that is given through the examples.
- In formatting conventions, emphasize that indentation is meant to be
  for alignment.
- Omitted a didactic section on roles and directives;
  instead I show them briefly in an example
- show section header syntax in action with an example
- Clarified code object references.

For DM-4001.