Create Git LFS guide
This document centralizes Git LFS user documentation that has otherwise
been stored in the READMEs of individual repos.

This is meant to be a top-level document in developer docs since our
implementation of Git LFS is unique and it is a developer-facing tool.

Builds on writing/edits from @jmatt and @frossie.

For DM-4469.