Use orphan for reST sample pages
For reST samples that are meant to be viewable *pages* we can use the
orphan field to tell Sphinx its okay they're not in the toctree.

In the code example in the sphinx doc the orphan field is omitted by
using the lines field of literalinclude. This saves us from users
accidentally putting the orphan field in their own docs (which should
generally never happen).

This is different from the development/docs/snippets/ since those were
never meant to be standalone pages, only includes.

For DM-4511.

Use lines field to hide the orphan field from samples