Design tree structure for release notes
The release directory looks like this:

releases
├── index.rst
├── v11_0
│   ├── index.rst
│   ├── metrics.rst
│   └── notes.rst
└── v11_1
    ├── index.rst
    ├── metrics.rst
    └── notes.rst

Each index.rst provides a toctree for Sphinx. In some ways it's a lot of
hierarchy. The notes.rst and metrics.rst documents could be combined,
but on Confluence the Characterization Metrics and Release notes are
setup as separate pages.

To help the user get to the current release note, I provide a direct
link from the main page. Going through this link bypasses see the
metrics page for that release. Perhaps the release notes could at least
link to the associated metrics document if it exists.

For DM-4536.