Port Python Style Guide from Confluence
Source:
https://confluence.lsstcorp.org/display/LDMDG/Python+Coding+Standard

In porting the content, I've standardized the hearing hierarchy to be

- Section
  - Rule 1

Each rule is now a sub-section that is labeled.

I've used intersphinx to link to Python built-ins as much as possible.

Note that the docstring section will need to be amended to point to the
new numpydoc standard.

For DM-4656.