Refactor Git and JIRA tips out of workflow doc
This work moves the Git Setup and JIRA Usage appendices out of the
Developer Workflow document and into the Developer Tools section of the
Developer Guide. This way the workflow guide is restricted to
requirements, while the Git and JIRA tip pages get more visibility to
readers browsing the contents page.

Through 'See also' links ensure that these docs are discoverable when a
reader is viewing a related document.

For DM-4793.