Port Using C++ Templates page from Confluence
This page demonstrates a new type of content that can live in the Coding
Guides section: coding tutorials (as opposed to style requirements).

Original content is from
https://confluence.lsstcorp.org/pages/viewpage.action?pageId=20284190
with example code properly refactored from the doc prose.

I think the article could use a re-write, but that will require
collaboration with a C++ developer.

For DM-4656.