Increase copy-and-pastability of Git LFS guide
This draft of the LFS docs reduce the chatiness and instead present the
reader with simple blocks that can be copied and pasted into their
config files.

Rather than having the reader interact with the same config file
multiple times, the instructions are refactored to have just one edit to
the config files per auth option branch. To keep the config comonalities
of read-only and read-write in sync, I've placed the config content in
snippet files that are referenced with a literalinclude. The lines field
ensures that the correct content is shown to both types of users.

For DM-4888.