Fix orphan warnings from rst style guide includes
The trick is to use `:orphan: True` to tell Sphinx that a
reStructuredText sample file shouldn't be included in the toctree.

But we don't want these orphan fields to appear in the included content.
For literalinclude directives (that show the source of the file), use
the :lines: field to skip the 'orphan' lines. For include directives
(that effectively insert that file's content into the rendered
reStructuredText source), use the start-line field. Note that start-line
is 0-based, while lines is 1-based.