Document lsst.utils.tests.MemoryTestCase
States that MemoryTestCase should always be used, and includes an
example of a test runner.

This documentation is added to the unit test policy document, although
it seems to point out the need for more practical, example-based,
documentation of our coding patterns.

When software documentation is available this doc should link to the
lsst.utils.tests API reference page.

See also:
https://community.lsst.org/t/what-is-the-policy-for-using-lsst-utils-tests-memorytestcase

For DM-5156.