Clarify division of Git setup and workflow docs
This commit solves the issue where a reader might go to git_setup.rst to
guidance on how to *use* Git for DM work rather than the official
workflow document (workflow.rst). These need to be separate pages
because the latter is a requirements document, and the former is intended to be
simply helpful.

- Rename "Git Setup and Best Practices" to "Git Configuration Best
  Practices" to make scope clearer.
- Expand the See also section to make it easier for someone who landed
  on the Git Configuration Best Practices document to make their way
  into the appropriate part of the Workflow document.