Modernize idiomatic python guidelines
Some of the existing idiomatic python guidelines were written from the
point of view of migrating from very old versions of Python to newer
versions. This commit attempts to rephrase some of the idiomatic python
guidelines to assume that we have a working Python 2.7 codebase and are
now porting to Python 3.

- Remove old exception handling guidance
- Remove discussion of apply that since it has been deprecated since
  2004.
- Fix up file/open guidance
- Cleanup section on subprocess
- Link to Python's lambda docs
- Clarify iters with python 3
- Various language cleanups and corrections.