Add link to edition dashboard below edition name
I had to hack the HTML and CSS a bit to get the same styling for the
edition link as for toctree links in the sidebar.

Note I'm using inline <style> just because we may deprecate this theme
in the near future, and don't care too much about building up a
fully-fledged custom style sheet.