More conservative on project-level vs DM-level
Due to discussion in LDM-401 (https://ls.st/std), LPM-51 is current
inconsistent in its language about the CCB that controls LDM documents
(likely an under-described edge case?). Because of this, I'm going to
release this document entirely from the task of assigning any handle to
any CCB. All I need is define the concept of a DM-level document and a
Porject level document for the procedure itself.