Integrate docstring placement+reference sections
Rather than separately documenting the placement of different types of
docstrings, we're now including this info with the reference sections
for each type of docstring. This is a less confusing approach because
folks expect to, for example, see everything about class docstrings in
one place.

Now the type-specific docstring reference sections are formatted as

1. Section listing
2. Placement info
3. Examples