Create /jenkins/ and LDF sections
This creates a Services section to complement the Development section.
Services focuses on our hosted and third-party infrastructure like
Jenkins, JIRA, GitHub, and of course NCSA services. These topics are
user guides and policies on how we use and govern these services.

This commit re-organizes the Jenkins CI topic into a new /jenkins/
section. The existing NCSA services remain in the /services/ directory
because there's little benefit making that directory name more specific
at this point and necessitating a large number of redirects.