Clarify how to use the sqre/infra/documenteer job
Use an example-based approach with a screenshot to make it easier to
understand how to run a ticket branch preview build with the
sqre/infra/documenteer job.

Also clarify that the "slug" field is the input for LSST the Docs to
determine the slug and that LTD uses that value to determine the
appropriate slug.