Change Struct/dict documentation to use def list
Each attribute of the struct is a definition list item. The attribute
name is the "term" and the description is the "definition." This is more
semantically meaningful than listing the struct's attributes.

This also adds documentation on dict without known keys, which is the
primary scenario for many dict parameters.

Finally, this also consolidates all documentation on documenting
parameters/return variables to the Parameters section. The Returns and
Yields sections link to this documentation.

Clarify formatting of dict keys