Remove pyvenv step from documenteer install
The pyenv strategy for isolating documenteer from the Stack's conda
environment seems to be error prone. It's easy to use the wrong Python
and end up with a Python environment that doesn't make sense. With
this commit we're removing mention of the virtual environment and
instead recommending that the documenteer package is installed directly
into the stack's conda environment. This emulates the likely endgame
anyhow, where documenteer is packaged with conda and distributed as part
of the conda environment.