Use pipelines.lsst.io weekly for intersphinx
Agreed on Slack with K-T that the weekly is a good balance of stability
and being able to link to relatively new APIs.

The 'pipelines' name for intersphinx is used to label "ref" links and
doesn't affect code cross links. "pipelines" is better than "lsst" here
because many different distributions share the "lsst" Python namespace.