Recommend anonymous inline links
The example of inline links was creating a reference (using a
single underscore), but some of our documents were using two
underscores to create an anonymous link.

The anonymous link seems more appropriate for the inline style.
If the link is referenced multiple times, the first syntax with
an explicit, separate URL seems better.  Recommend this and
adjust the inline example to use an anonymous link.