add more python unit test decorators
The classParameters and methodParameters decorators are useful,
but sometimes you want to iterate over all combinations of some
parameters, but don't want to go to the trouble of listing all
combinations. This commit adds classParametersProduct and
methodParmaetersProduct that generates the cartesian product
of the provided parameters, and iterates over them.